Update and fixed OpenVPN GUI Spanish translation
In addition to translating what was untranslated, the most important fixes are in password-related concepts. In many areas the "password" concept was translated as "clave", which is used in colloquial SPanish but it shouldn't be used when that translation could be confused with "key", which in Spanish regarding certificates also is translated as "clave". To avoid this, I translated the "password" concept as "contraseña" in all cases, like seen in Windows and all other major desktop, mobile and web platforms/apps.